# perf(home): read Active Players from game_recent_players table