This PR adjusts `LoadThinActivePlayersListAction` (the home page Active Players query) to read from the new `game_recent_players` table rather than `UserAccounts` for a significant query performance improvement.

**Benchmark Results (100 iterations):**

| Implementation | Average Time | Min Time | Max Time |
|----------------|--------------|----------|----------|
| Old (UserAccounts) | 206.95 ms | 201.33 ms | 218.75 ms |
| New (game_recent_players) | 90.92 ms | 88.28 ms | 109.19 ms |

The new query maintains the same data contract used by the old one. The SWR cache is retained, so it's probably not much of a difference in the grand scheme of things. I just wanted the cache miss to be a bit faster.